requirements.txt
Pillow
opencv-python
xmltodict